implements week2-4 assignment

+++ week2/week2-1.sol
@@ -5,7 +5,7 @@ import "./IERC20.sol";
 
 contract ERC20_Week2_1 is IERC20 {
 
-    uint public totalSupply;
+    uint256 public totalSupply;
     mapping(address => uint) public balanceOf;
     mapping(address => mapping(address => uint)) public allowance;
     string public name = "Test";
